requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
babel==2.17.0
base58==2.1.1
bitarray==3.4.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chatgpt_md_converter==0.3.6
click==8.2.1
colorama==0.4.6
construct==2.10.68
construct-typing==0.6.2
crc16==0.1.1
crc32c==2.7.1
distro==1.9.0
fastapi==0.115.12
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.10
frozenlist==1.6.0
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonalias==0.1.1
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
mcp==1.9.2
multidict==6.4.3
openai==1.82.1
openai-agents==0.0.16
ordered-set==4.1.0
propcache==0.3.1
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytonapi==0.4.9
pytonlib==0.0.65
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
solana==0.36.7
solders==0.26.0
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tqdm==4.67.1
tvm-valuetypes==0.0.12
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==2.3.1
websockets==15.0.1
yarl==1.20.0
